feat: make coverTransactionCost standalone component
